move udp server to test file, add link to BEP15
